Extract search controller for reuse
